+++ packages/sdk/react-universal/src/index.test.ts
@@ -1,0 +1,3 @@
+describe('react-universal', () => {
+  test.todo('Add react universal sdk tests');
+});