+++ packages/store/node-server-sdk-dynamodb/CHANGELOG.md
@@ -90,6 +90,17 @@
   * devDependencies
     * @launchdarkly/node-server-sdk bumped from 9.2.1 to 9.2.2
 
+## [6.1.16](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.1.15...node-server-sdk-dynamodb-v6.1.16) (2024-07-01)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/node-server-sdk bumped from 9.4.6 to 9.4.7
+  * peerDependencies
+    * @launchdarkly/node-server-sdk bumped from >=9.4.3 to >=9.4.7
+
 ## [6.1.15](https://github.com/launchdarkly/js-core/compare/node-server-sdk-dynamodb-v6.1.14...node-server-sdk-dynamodb-v6.1.15) (2024-06-14)